yolov4/eval_code/requirements.txt
pillow
torch
torchvision
numpy
scipy
scikit-image
tqdm
